Move source files into their own Go packages
